fix unsigned int problem
Former-commit-id: aa4e0be4efe8e818b711a720b13005c272a3e9fc

+++ GeoImage/src/main/java/org/geoimage/impl/envi/EnvisatImage.java
@@ -335,7 +335,7 @@ public class EnvisatImage extends SarImageReader {
 
     // </editor-fold>
     @Override
-    public int readPixel(int x, int y, int band) {
+    public long readPixel(int x, int y, int band) {
         int result = 0;
         long temp = 0;
         byte[] pixelByte = new byte[2];
